requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#

# Core Data Processing
pyspark>=3.4.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Configuration Management
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Data Pipeline
requests>=2.31.0
apache-airflow>=2.7.0

# Only add Apache Spark provider if not in base image
# apache-airflow-providers-apache-spark>=4.11.0

# Note: apache-airflow and all other dependencies are already 
# in the apache/airflow:2.10.0-python3.12 base image
